Apply modularizations and simplification
