+++ client/utils/api.js
@@ -1,10 +1,33 @@
+function parseJSON(text) {
+  if (!text) return null;
+
+  try {
+    return JSON.parse(text);
+  } catch {
+    return null;
+  }
+}
+
+export async function responseErrorMessage(res) {
+  const text = await res.text();
+  const data = parseJSON(text);
+  const fallback = `Request failed (${res.status})`;
+
+  return data?.error || data?.message || (text.trim() && !text.trim().startsWith('<') ? text.trim() : fallback);
+}
+
 export async function fetchJSON(url, options = {}) {
   const res = await fetch(url, options);
   const text = await res.text();
-  const data = text ? JSON.parse(text) : null;
+  const data = parseJSON(text);
 
   if (!res.ok) {
-    throw new Error(data?.error || data?.message || `Request failed (${res.status})`);
+    const fallback = `Request failed (${res.status})`;
+    throw new Error(data?.error || data?.message || (text.trim() && !text.trim().startsWith('<') ? text.trim() : fallback));
+  }
+
+  if (text && data === null) {
+    throw new Error('Invalid JSON response');
   }
 
   return data;
